Slightly improve quality of test failure messages
This uses `assert_eq!` with `None`, instead of `assert!` on an
`is_none()` call, so that we get `left` and `right` expressions
in test failure output.

It also includes some other changes that are just for code
readability: running `cargo fmt`, and replacing an expression in
the tests that evaluates to the same string as `crate::NULL_DEVICE`
with that (as the goal is not to test the value of `NULL_DEVICE`).