Remove `configure_command_msys*` tests at least for now
Because they are not really testing the right thing, due to the
distinction between inherited and `.env()`-specified environment.

See #1578.

While for `configure_command_clears_external_config` this slightly
diminishes but largely leaves intact the value of the test, the
`configure_command_msys` and `configure_command_msys_extends` tests
are really verifying almost nothing, and the latter (for this
reason) cannot pass.

A variant of it for testing the preceding fix to append to
arbitrary environment variable values would likewise not be able to
pass. So that is not added at this time.

Because the `configure_command_clears_external_config` test remains
useful, it is not removed.