Start testing that installation config is not from `GIT_CONFIG`
These tests should not pass yet, because we currently do allow
`GIT_CONFIG` to be read. Although the optimization that reads
`EXEPATH` will sometimes prevent it from being used, it will
usually be used, becuase `EXEPATH` is usually not set on Windows
(and outside of Windows, that optimization is inapplicable).
Furthermore, the tests in `gix-path/src/env/git/tests.rs` bypass
that optimization.

Two of the three new tests rightly fail. But one of them,
`never_from_git_config_env_var`, wrongly passes, due to a big in
the test, described in the comments there.