Don't test `jj_realistic_needs_to_be_more_clever` on Windows
This conditionally marks the `jj_realistic_needs_to_be_more_clever`
test as ignored, when the target is Windows. But the test is still
built on Windows, and can be run with:

    cargo nextest run -p gix -- --ignored jj_realistic_needs_to_be_more_clever

This is a workaround for #1575.

When run on Windows, the test previously checked if it was running
on CI and skipped its contents. This removes that logic, since the
test no longer runs by default on CI or otherwise.