feat: Recognize `GIX_TEST_CREATE_ARCHIVES_EVEN_ON_CI`
When `gix-testtools` runs fixtures, it now recognizes a new
environment variable, `GIX_TEST_CREATE_ARCHIVES_EVEN_ON_CI`,
specifying that archives should be generated even when on CI.

By default, they are still not generated when on CI. It may make
sense to enable this:

- If automatically testing archive creation, or

- As a way to check that all intended generated arhives are committed
  (which is the motivating use case for this feature), or

- If actually using CI to generate archives that will be uploaded
  as artifacts, or

- In unusual non-CI environments that are mis-detected as CI
  (though that should usually be investigated and fixed, since some
  software performs destructive operations more readily without
  interactive checks when CI is detected).

The usual reason for not generating archives on CI is that they
would not typically be preserved. Thus refraining from generating
them on CI remains the default behavior.

Like the `GIX_TEST_IGNORE_ARCHIVES` environment variable, the new
variable `GIX_TEST_CREATE_ARCHIVES_EVEN_ON_CI` is currently
interpreted as "true" based solely on its presence. This is to say
that is actual value is currently not examined.