Note that how we set `MSYS` ignores `env`/`env_remove` calls
This problem can be avoided by a design that doesn't create the
assumption that such calls would have an effect, or that prevents
them from being made (e.g. a builder funtion), or by accounting for
and honoring them (which may involve reimplementing a fragment of
the Windows API based case-folding that `std::process::Command`
uses). But none of these approaches really seems overall better
than what we are doing now. So for now, keep the design but note
this one confusing spot.

This is a private function, so I've just commented by where the
relevant logic is, rather than using a `///` comment.