Rename nonexistent repo in `Makefile` from `foo` to `nonexistent`
The nonexistent repo used to get data from a 401 response in the
`update-curl-fixtures` rule was formerly called `foo`. (In 1ea08bf
in #1624, I had erroneously said this was for a 404 response, but
the `foo` URL is the 401 URL, not the 404 URL.) This commit renames
it to `nonexistent`, which is more self-documenting and less likely
to be created by accident, even temporarily.

In 1ea08bf (#1624), it was changed from `Byron/foo` to
`GitoxideLabs/foo`, going along with the change of the existing
repository (gitoxide's own) from `Byron/gitoxide` to
`GitoxideLabs/gitoxide`. But while `Byron/foo` would continue not
to exist as long as the convention of not having a `foo` repo was
maintained by `@Byron`, a `foo` repository in `GitoxideLabs` might
someday be created by another developer who is not aware of its
status in `Makefile`. Renaming it to `nonexistent` should make this
easier to avoid.

I tested this change with the same approach as noted for 1ea08bf.