Only run expectrl gix-prompt tests where supported
Two important `gix-prompt` tests simulate interactivity with
`expectrl`, which (non-optionally) depends on `ptyprocess`. But the
only (families of) Unix-like operating systems that `ptyprocess`
currently supports are Linux, FreeBSD, and macOS.

The tests that used `expectrl` were previously targeted to all
"unix" systems. This narrows them to systems where `target_os` is
`linux`, `freebsd`, or `macos`, and conditions the `expectrl` dev
dependency on those targets.

This way, running the test suite no longer fails to build on other
Unix-like systems. (In particular, this fixes the failure on
OmniOS, an illumos system. But the problem isn't illumos-specific.)