Avoid unneeded +x in make_rev_spec_parse_repos
The make_rev_spec_parse_repos fixture script used `chmod 755` to
make a looose object file writable (so it could be corrupted, to
test the treatment of corrupted objects). But this also made it
executable, which was not needed (on any platforms).

This commit changes it from 755 to 644 so that it is still made
writable, but not made executable. (It does not start out being
executable.) This does not affect the results of any tests.