Use `--workspace` where `--all` is deprecated
Most `cargo` commands that support `--all` to affect the entire
workspace rather than a single package have deprecated it in favor
of `--workspace` and treat them as synonyms. The major exception
(at least as relevant to this repository) is `cargo fmt`, which
recognizes `--all` and not `--workspace`.

This replaces `--all` with `--workspace` in `cargo check`,
`cargo doc`, and `cargo nextest` commands in a CI workflow and the
`justfile`. All these subcommands have deprecated `--all` in favor
of `--workspace`, and document this in relevant `--help` output.

(Other major subcommands that recommend `--workspace` over its
deprecated `--all` alias include `cargo build` and `cargo test`,
but there are currently no occurrences of such commands with
`--all` in this repository.)