Mark gix-pack vendored `izip!` doctest `ignore`
That macro is not exported, and adjusting the doctest so that it
can run successfully (or even compile) might be tricky without a
significant change (or adding dependencies). This fixes a failure
when running doctests for the whole workspace (as CI now includes).

This change is not "user-facing", as the macro is not part of the
public interface of `gix_pack`. Its documentation comment doesn't
appear https://docs.rs/gix-pack/0.54.0/gix_pack/index/index.html.
Nonetheless, this adds a note to clarify the change, to avoid
attributing it to the original code from which this was vendored.