Make YAML style clearer and more consistent
In GitHub Actions workflow files, this separates top-level keys
within a job with blank lines, which was sometimes but not usually
already done, since this seems to improve readability. Besides
that, this also applies the prevailing style more consistently.

This style tweak is in preparation for adding explicit
`permissions` keys (so that doing so won't decrease readability).