Temporarily keep i686 job failure from canceling armv7 job
`test-32bit` fails in the job for the `i686-unknown-linux-musl`
target since changing `-p gix-hashtable` to `-p gix` in the test
step because various data structure size assertions fail, which is
because some data structures are smaller on 32-bit archictectures
than on 64-bit architectures.

That is unrelated to the commented problem of environment not being
passed into the test environment, which does not happen on
`i686-unknown-linux-musl`. But maybe it would happen on
`armv7-linux-androideabi`, due to its use of cross.

The i686 job, which doesn't require emulation, is faster, and when
it fails due to those assertions, it keeps the armv7 job from
getting far enough to reveal if it has its own environment-related
(or other) failures. This temporarily sets `fail-fast: false` so
that the armv7 job can get far enough that we find out.