Change quoted nextest `--all` arg to `--workspace`
This changes the deprecated `--all` option for `cargo nextest` to
the recommended (and equivelent) `--workspace` option, in quoted
default argument values for recipes in `justfile`.

Other occurrences of `--all` to `cargo nextest`, and other commands
for which `--all` is a deprecated alias of `--workspace`, were
changed to `--workspace` in 14d472d (#1654), but I had missed
these.