Don't use `with_program find` in journey tests
This removes all uses of `with_program`, which is no longer used
with any command other than `find`. The rationale is:

- `find` is a standard command. Minimal systems may omit it, but
  they omit numerous other software that is needed for building and
  testing gitoxide. `find` is standardized and required by POSIX
  and is present on all common general-purpose installations of
  Unix-like operating systems. It is also present in the Git Bash
  environment (and most other MSYS2 environments) that may be used
  for running the test suite on Windows.

- One purpose of `with_program` is to check for missing commands
  where, if not checked for, then tests might run but wrongly pass
  due to assertions that commands fail (where they should fail for
  a different reason). However, that does not seem to apply to any
  `with_program find`: all assertions with `find` in code to which
  that had been applied are using `expect_run_sh $SUCCESSFULLY`.

`with_program` is typically applied in a `(` `)` subshell, which
itself can have an important effect in isolating changes made
within it. In this case, however, the code was either still
enclosed equivalently, or appeared at the end of a larger `(` `)`
scope. So this removes `(` `)` subshells that only existed to
facilitate the deleted `with_program` uses and dedents.