Try to also log to terminal, interleaved with cargo output
So the `cc1` calls can also be seen in context.

`/dev/tty` is used instead of writing to standard output because
standard output is expectd to be redirected and may in some cases
even be consulted by code in a build script, whereas we really do
want to write to the terminal.

This also:

- Has the `wrap1` script use `-e` so that it will stop at the first
  error. No command before the last one should fail. If the `flock`
  command fails, either due to a lock-related error or due to an
  error in `tee`, we want to know about it.

  I should've done this before, but it's more important now that
  I'm also tee-ing to `/dev/tty`: it's not obvious that this would
  work on CI.

- Tweaks the code style of the `printf` command, to make the
  command slightly more readable, and make clearer that this is
  really just being used as a portable alternative to `echo`.